+++ cvat-ui/src/actions/tasks-actions.ts
@@ -90,10 +90,12 @@ ThunkAction<Promise<void>, {}, {}, AnyAction> {
             return;
         }
 
-        const array = Array.from(result);
+        const array = Array.from(result)
+            .filter((task: any) => task.jobs.length > 0);
         const previews = [];
         const promises = array
             .map((task): string => (task as any).frames.preview());
+
         dispatch(
             getInferenceStatusAsync(
                 array.map(
@@ -441,23 +443,23 @@ function updateTask(): AnyAction {
     return action;
 }
 
-function updateTaskSuccess(taskInstance: any): AnyAction {
+function updateTaskSuccess(task: any): AnyAction {
     const action = {
         type: TasksActionTypes.UPDATE_TASK_SUCCESS,
         payload: {
-            taskInstance,
+            task,
         },
     };
 
     return action;
 }
 
-function updateTaskFailed(error: any, taskInstance: any): AnyAction {
+function updateTaskFailed(error: any, task: any): AnyAction {
     const action = {
         type: TasksActionTypes.UPDATE_TASK_FAILED,
         payload: {
             error,
-            taskInstance,
+            task,
         },
     };