+++ cvat-ui/src/utils/plugin-checker.ts
@@ -7,34 +7,32 @@ const core = getCore();
 class PluginChecker {
     public static async check(plugin: SupportedPlugins): Promise<boolean> {
         const serverHost = core.config.backendAPI.slice(0, -7);
-        const isReachable = async (url: string): Promise<boolean> => {
+        const isReachable = async (url: string, method: string): Promise<boolean> => {
             try {
-                await core.server.request(url);
+                await core.server.request(url, {
+                    method,
+                });
                 return true;
             } catch (error) {
-                if (error.code === 404) {
-                    return false;
-                }
-
-                throw error;
+                return ![0, 404].includes(error.code);
             }
         };
 
         switch (plugin) {
             case SupportedPlugins.GIT_INTEGRATION: {
-                return isReachable(`${serverHost}/git/repository/meta/get`);
+                return isReachable(`${serverHost}/git/repository/meta/get`, 'OPTIONS');
             }
             case SupportedPlugins.AUTO_ANNOTATION: {
-                return isReachable(`${serverHost}/auto_annotation/meta/get`);
+                return isReachable(`${serverHost}/auto_annotation/meta/get`, 'OPTIONS');
             }
             case SupportedPlugins.TF_ANNOTATION: {
-                return isReachable(`${serverHost}/tensorflow/annotation/meta/get`);
+                return isReachable(`${serverHost}/tensorflow/annotation/meta/get`, 'OPTIONS');
             }
             case SupportedPlugins.TF_SEGMENTATION: {
-                return isReachable(`${serverHost}/tensorflow/segmentation/meta/get`);
+                return isReachable(`${serverHost}/tensorflow/segmentation/meta/get`, 'OPTIONS');
             }
             case SupportedPlugins.ANALYTICS: {
-                return isReachable(`${serverHost}/analytics/app/kibana`);
+                return isReachable(`${serverHost}/analytics/app/kibana`, 'GET');
             }
             default:
                 return false;