+++ cvat/apps/auto_segmentation/views.py
@@ -1,9 +1,11 @@
 
-# Copyright (C) 2018 Intel Corporation
+# Copyright (C) 2018-2019 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
+
 from django.http import HttpResponse, JsonResponse, HttpResponseBadRequest
+from rest_framework.decorators import api_view
 from rules.contrib.views import permission_required, objectgetter
 from cvat.apps.authentication.decorators import login_required
 from cvat.apps.engine.models import Task as TaskModel
@@ -185,9 +187,10 @@ def create_thread(tid, labels_mapping, user):
         try:
             slogger.task[tid].exception('exception was occured during auto segmentation of the task', exc_info=True)
         except Exception:
-            slogger.glob.exception('exception was occured during auto segmentation of the task {}'.format(tid), exc_into=True)
+            slogger.glob.exception('exception was occured during auto segmentation of the task {}'.format(tid), exc_info=True)
         raise ex
 
+@api_view(['POST'])
 @login_required
 def get_meta_info(request):
     try:
@@ -204,7 +207,7 @@ def get_meta_info(request):
 
         return JsonResponse(result)
     except Exception as ex:
-        slogger.glob.exception('exception was occured during tf meta request', exc_into=True)
+        slogger.glob.exception('exception was occured during tf meta request', exc_info=True)
         return HttpResponseBadRequest(str(ex))