+++ cvat/apps/engine/views.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2018 Intel Corporation
+# Copyright (C) 2018-2019 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
@@ -11,8 +11,9 @@ import shutil
 from datetime import datetime
 from tempfile import mkstemp
 
-from django.http import HttpResponseBadRequest
-from django.shortcuts import redirect, render
+from django.views.generic import RedirectView
+from django.http import HttpResponseBadRequest, HttpResponseNotFound
+from django.shortcuts import render
 from django.conf import settings
 from sendfile import sendfile
 from rest_framework.permissions import IsAuthenticated
@@ -32,7 +33,6 @@ from django.utils import timezone
 from . import annotation, task, models
 from cvat.settings.base import JS_3RDPARTY, CSS_3RDPARTY
 from cvat.apps.authentication.decorators import login_required
-import logging
 from .log import slogger, clogger
 from cvat.apps.engine.models import StatusChoice, Task, Job, Plugin
 from cvat.apps.engine.serializers import (TaskSerializer, UserSerializer,
@@ -53,14 +53,22 @@ import cvat.apps.dataset_manager.task as DatumaroTask
 @login_required
 def dispatch_request(request):
     """An entry point to dispatch legacy requests"""
-    if request.method == 'GET' and 'id' in request.GET:
+    if 'dashboard' in request.path or (request.path == '/' and 'id' not in request.GET):
+        return RedirectView.as_view(
+            url=settings.UI_URL,
+            permanent=True,
+            query_string=True
+        )(request)
+    elif request.method == 'GET' and 'id' in request.GET and request.path == '/':
         return render(request, 'engine/annotation.html', {
             'css_3rdparty': CSS_3RDPARTY.get('engine', []),
             'js_3rdparty': JS_3RDPARTY.get('engine', []),
-            'status_list': [str(i) for i in StatusChoice]
+            'status_list': [str(i) for i in StatusChoice],
+            'ui_url': settings.UI_URL
         })
     else:
-        return redirect('/dashboard/')
+        return HttpResponseNotFound()
+
 
 class ServerViewSet(viewsets.ViewSet):
     serializer_class = None