+++ cvat/settings/base.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2018 Intel Corporation
+# Copyright (C) 2018-2019 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
@@ -30,8 +30,9 @@ INTERNAL_IPS = ['127.0.0.1']
 
 try:
     sys.path.append(BASE_DIR)
-    from keys.secret_key import SECRET_KEY
+    from keys.secret_key import SECRET_KEY # pylint: disable=unused-import
 except ImportError:
+
     from django.utils.crypto import get_random_string
     keys_dir = os.path.join(BASE_DIR, 'keys')
     if not os.path.isdir(keys_dir):
@@ -91,7 +92,6 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'cvat.apps.engine',
-    'cvat.apps.dashboard',
     'cvat.apps.authentication',
     'cvat.apps.documentation',
     'cvat.apps.git',
@@ -191,7 +191,10 @@ UI_HOST = os.environ.get('UI_HOST', 'localhost')
 UI_PORT = os.environ.get('UI_PORT', '3000')
 CORS_ALLOW_CREDENTIALS = True
 CSRF_TRUSTED_ORIGINS = [UI_HOST]
-UI_URL = '{}://{}:{}'.format(UI_SCHEME, UI_HOST, UI_PORT)
+UI_URL = '{}://{}'.format(UI_SCHEME, UI_HOST)
+if len(UI_URL):
+    UI_URL += ':{}'.format(UI_PORT)
+
 CORS_ORIGIN_WHITELIST = [UI_URL]
 CORS_REPLACE_HTTPS_REFERER = True